+++ backend/testing/test_apis.py
@@ -1,10 +1,7 @@
 import json
-from unicodedata import category
 from django.urls import reverse
-from django.db.models import Sum
 from django.utils import timezone
 from rest_framework.test import APITestCase, APIRequestFactory
-from rest_framework.views import APIView
 from kalunwa.content.serializers import StatusEnum
 from .utils import  (
     ABOUT_US_CAMP_URL, ABOUT_US_LEADERS, ABOUT_US_TOTAL_MEMBERS,
@@ -14,8 +11,17 @@ from .utils import  (
     to_formal_mdy, HOMEPAGE_JUMBOTRON_URL, HOMEPAGE_EVENT_URL
 )
 from kalunwa.content.models import(
-    CampEnum, CampLeader, CampPage, Contributor, Demographics,  Image, Jumbotron, 
-    News, OrgLeader, Project, Event
+    CampEnum, 
+    CampLeader, 
+    CampPage, 
+    Contributor, 
+    Demographics,  
+    Image, 
+    Jumbotron, 
+    News, 
+    OrgLeader, 
+    Project, 
+    Event
 )
 from rest_framework import status
 #-------------------------------------------------------------------------------
@@ -935,78 +941,6 @@ class ProjectGetTestCase(APITestCase):
             'category' : expected_contributor.category.label
         } 
         self.assertDictEqual(response_contributor[0], expected_contributor_data)
-
-
-class QueryLimitTestCase(APITestCase):
-    """
-    mock viewset that uses this logic e.g. Event
-    -> test on list endpoint
-        - mock 5 events
-        - query limit is an integer.
-        - query limit values to test: [-1, 0, 3, 5, 6]
-            # negative value
-            # zero
-            # less than total events
-            # exact no. of events
-            # greater than no. of events
-            # strings
-                # empty string
-    """
-    @classmethod
-    def setUpTestData(cls) -> None:
-        cls.image_file = get_test_image_file()
-
-        for _ in range(5): 
-            Event.objects.create(
-            title= f'Event {_}', 
-            description= f'description {_}',
-            start_date=timezone.now(),
-            end_date=timezone.now(),
-            image = Image.objects.create(name=f'image_{_}', image=cls.image_file),  
-            is_featured=True,
-            )           
-
-        cls.event_count = len(Event.objects.all())            
-
-    def test_expected_query_integer_input(self):
-        # 0 -> returns 0 or no events
-        query_limit = 0
-        response = self.client.get(f'/api/events/?query_limit={query_limit}')        
-        self.assertEqual(len(response.data), 0)
-        # 3 -> returns 3 events
-        query_limit = 3
-        response = self.client.get(f'/api/events/?query_limit={query_limit}')        
-        self.assertEqual(len(response.data), query_limit)        
-        # 5 -> returns 5 events
-        query_limit = 5
-        response = self.client.get(f'/api/events/?query_limit={query_limit}')        
-        self.assertEqual(len(response.data), query_limit)           
-        # 6 -> returns 5 events
-        query_limit = 6
-        response = self.client.get(f'/api/events/?query_limit={query_limit}')        
-        self.assertEqual(len(response.data), self.event_count)
-        # aaa -> strings are ignored
-
-    def test_negative_query_integer_input(self):
-        # -1 -> ignores negative, return all events
-        query_limit = -1
-        response = self.client.get(f'/api/events/?query_limit={query_limit}')
-        self.assertEqual(len(response.data), self.event_count)
-
-    def test_string_query_input(self):
-        query_limits = ['aaa', '*&()', '']
-
-        for query_limit in query_limits:
-            response = self.client.get(f'/api/events/?query_limit={query_limit}')        
-            self.assertEqual(len(response.data), self.event_count)   
-
-"""
-Test QueryLimitBackend: Gallery
-
-tests:
-    - if no model -> error 
-    - is used by a model that has no gallery 
-""" 
 
 
 # ---------------------------------------------------------------------------
